requirements.txt
altair==5.4.1
asgiref==3.8.1
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.1
flatbuffers==24.3.25
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
mysql-connector-python==9.0.0
narwhals==1.8.4
numpy==2.0.1
packaging==24.1
pandas==2.2.3
pillow==10.4.0
protobuf==5.27.3
pyarrow==17.0.0
PyAutoGUI==0.9.54
pydeck==0.9.1
pygame==2.6.0
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyreadline3==3.4.1
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
sqlparse==0.5.1
streamlit==1.39.0
sympy==1.13.2
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.2
